# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths 
basePath: /jsonata
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
tags:
  - name: Info
    description: Informational methods.
  - name: Spec
    description: Specification related endpoints (CRUD)
    
paths:
  /info/ping:
    x-swagger-router-controller: infoController
    get:
      tags:
        - Info
      description: Returns an acknowledgement that the service is up and running
      operationId: ping
      responses:
        "200":
          description: Success
          schema:
            type: string
          
  /info/about:
    x-swagger-router-controller: infoController
    get:
      tags:
        - Info
      description: Returns information about this service.
      operationId: about
      responses:
        "200":
          description: Success
          schema:
            type: string
  
  /spec/{specName}:
    x-swagger-router-controller: specificationController
    parameters:
      - name: specName
        in: path
        description: the identifier fo the specification being requested
        required: true
        type: string
    get:
      tags:
        - Spec
      description: Retrieves a given spec provided by the specName
      operationId: getSpec
      responses:
        "200":
          description: success
          schema:
            type: string
    post:
      tags:
        - Spec
      consumes:
        - text/plain
      description: Saves a new spec under specName
      operationId: saveSpec
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: string
      responses:
        "200":
            description: success
            schema:
              type: string          
  /transform/{specName}:
    x-swagger-router-controller: transformController
    parameters:
      - name: specName
        in: path
        description: the identifier fo the specification being requested
        required: true
        type: string
    post:
      description: Transforms the provided body with the spec defined by specName
      operationId: transform
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The new JSON transformed by the spec
          schema:
            type: string

  /hello:
    # binds a127 app logic to a route
    x-swagger-router-controller: hello_world
    get:
      description: Returns 'Hello' to the caller
      # used as the method name of the controller
      operationId: hello
      parameters:
        - name: name
          in: query
          description: The name of the person to whom to say hello
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
    
  
  
# complex objects have schema definitions
definitions:
  HelloWorldResponse:
    required:
      - message
    properties:
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  GetMoviesListResponse:
    required:
      - movies
    properties:
      # The array of movies
      movies:
        type: array
        items: 
          type: object
          properties:
            id:
              type: string
            title:
              type: string
            year:
              type: number